Moved countries to constants.
